extract cascade module
